# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  completed: boolean;
  priority: TaskPriority;
  deadline?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskRequest {
  title: string;
  description?: string | null;
  priority: CreateTaskRequestPriority;
  deadline?: string | null;
}

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  completed?: boolean;
  priority?: UpdateTaskRequestPriority;
  deadline?: string | null;
}

export interface Expense {
  id: number;
  title: string;
  amount: number;
  category: string;
  date: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateExpenseRequest {
  title: string;
  amount: number;
  category: string;
  date: string;
  notes?: string | null;
}

export type SubscriptionBillingCycle =
  (typeof SubscriptionBillingCycle)[keyof typeof SubscriptionBillingCycle];

export const SubscriptionBillingCycle = {
  monthly: "monthly",
  quarterly: "quarterly",
  yearly: "yearly",
} as const;

export interface Subscription {
  id: number;
  name: string;
  amount: number;
  billingCycle: SubscriptionBillingCycle;
  category: string;
  nextBillingDate: string;
  isActive: boolean;
  notes?: string | null;
  createdAt: string;
}

export type CreateSubscriptionRequestBillingCycle =
  (typeof CreateSubscriptionRequestBillingCycle)[keyof typeof CreateSubscriptionRequestBillingCycle];

export const CreateSubscriptionRequestBillingCycle = {
  monthly: "monthly",
  quarterly: "quarterly",
  yearly: "yearly",
} as const;

export interface CreateSubscriptionRequest {
  name: string;
  amount: number;
  billingCycle: CreateSubscriptionRequestBillingCycle;
  category: string;
  nextBillingDate: string;
  notes?: string | null;
}

export type UpdateSubscriptionRequestBillingCycle =
  (typeof UpdateSubscriptionRequestBillingCycle)[keyof typeof UpdateSubscriptionRequestBillingCycle];

export const UpdateSubscriptionRequestBillingCycle = {
  monthly: "monthly",
  quarterly: "quarterly",
  yearly: "yearly",
} as const;

export interface UpdateSubscriptionRequest {
  name?: string;
  amount?: number;
  billingCycle?: UpdateSubscriptionRequestBillingCycle;
  category?: string;
  nextBillingDate?: string;
  isActive?: boolean;
  notes?: string | null;
}