# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Productivity Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management
  - name: expenses
    description: Expense tracking
  - name: subscriptions
    description: Subscription management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted

  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get all expenses
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Expense deleted

  /subscriptions:
    get:
      operationId: getSubscriptions
      tags: [subscriptions]
      summary: Get all subscriptions
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a new subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  /subscriptions/{id}:
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionRequest"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Delete a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Subscription deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        deadline:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - completed
        - priority
        - createdAt
        - updatedAt

    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        deadline:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - priority

    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        deadline:
          type: string
          format: date-time
          nullable: true

    Expense:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - amount
        - category
        - date
        - createdAt

    CreateExpenseRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required:
        - title
        - amount
        - category
        - date

    Subscription:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, yearly]
        category:
          type: string
        nextBillingDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - amount
        - billingCycle
        - category
        - nextBillingDate
        - isActive
        - createdAt

    CreateSubscriptionRequest:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, yearly]
        category:
          type: string
        nextBillingDate:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
      required:
        - name
        - amount
        - billingCycle
        - category
        - nextBillingDate

    UpdateSubscriptionRequest:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
        billingCycle:
          type: string
          enum: [monthly, quarterly, yearly]
        category:
          type: string
        nextBillingDate:
          type: string
          format: date-time
        isActive:
          type: boolean
        notes:
          type: string
          nullable: true