# lib/api-zod/src/generated/types/createExpenseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseRequest {
  title: string;
  amount: number;
  category: string;
  date: Date;
  notes?: string | null;
}