# lib/api-zod/src/generated/types/createSubscriptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSubscriptionRequestBillingCycle } from "./createSubscriptionRequestBillingCycle";

export interface CreateSubscriptionRequest {
  name: string;
  amount: number;
  billingCycle: CreateSubscriptionRequestBillingCycle;
  category: string;
  nextBillingDate: Date;
  notes?: string | null;
}