# lib/api-zod/src/generated/types/createTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaskRequestPriority } from "./createTaskRequestPriority";

export interface CreateTaskRequest {
  title: string;
  description?: string | null;
  priority: CreateTaskRequestPriority;
  deadline?: Date | null;
}