# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionBillingCycle } from "./subscriptionBillingCycle";

export interface Subscription {
  id: number;
  name: string;
  amount: number;
  billingCycle: SubscriptionBillingCycle;
  category: string;
  nextBillingDate: Date;
  isActive: boolean;
  notes?: string | null;
  createdAt: Date;
}