# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  completed: boolean;
  priority: TaskPriority;
  deadline?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}