# lib/api-zod/src/generated/types/updateSubscriptionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSubscriptionRequestBillingCycle } from "./updateSubscriptionRequestBillingCycle";

export interface UpdateSubscriptionRequest {
  name?: string;
  amount?: number;
  billingCycle?: UpdateSubscriptionRequestBillingCycle;
  category?: string;
  nextBillingDate?: Date;
  isActive?: boolean;
  notes?: string | null;
}