# lib/api-zod/src/generated/types/updateTaskRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Productivity Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskRequestPriority } from "./updateTaskRequestPriority";

export interface UpdateTaskRequest {
  title?: string;
  description?: string | null;
  completed?: boolean;
  priority?: UpdateTaskRequestPriority;
  deadline?: Date | null;
}